remove reference to "standalone" vs "multiservice" deployment, only support starter kit microservices for each GA4GH api going forward
