+++ src/main/java/org/ga4gh/starterkit/drs/app/package-info.java
@@ -2,9 +2,8 @@
  * Enables running the Starter Kit DRS as a standalone web service
  * <p>
  * This package contains the classes necessary to deploy the Starter Kit DRS
- * as a standalone web service, that is, with no other GA4GH APIs running as
- * part of the same application. Contains the main method, classes for loading
- * configurations, and Spring bean config for standalone deployments
+ * as a standalone web service. Contains the main method, classes for loading
+ * configurations, and Spring bean configuration
  * </p>
  * 
  * @since 0.1.4
