--- src/main/java/org/ga4gh/starterkit/drs/beanconfig/package-info.java
@@ -1,8 +1,0 @@
-/**
- * Contains Spring configurations that are to be used by the DRS API in all
- * contexts (ie as part of both standalone and non-standalone deployments)
- * 
- * @since 0.1.4
- * @version 0.1.4
- */
-package org.ga4gh.starterkit.drs.beanconfig;
