+++ src/test/java/org/ga4gh/starterkit/drs/controller/ObjectsTest.java
@@ -2,9 +2,8 @@ package org.ga4gh.starterkit.drs.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import org.ga4gh.starterkit.drs.app.DrsStandaloneServer;
-import org.ga4gh.starterkit.drs.app.DrsStandaloneSpringConfig;
-import org.ga4gh.starterkit.drs.beanconfig.StarterKitDrsSpringConfig;
+import org.ga4gh.starterkit.drs.app.DrsServer;
+import org.ga4gh.starterkit.drs.app.DrsServerSpringConfig;
 import static org.ga4gh.starterkit.drs.constant.DrsApiConstants.DRS_API_V1;
 import org.ga4gh.starterkit.drs.model.DrsObject;
 import org.ga4gh.starterkit.drs.testutils.ResourceLoader;
@@ -29,9 +28,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest
 @ContextConfiguration(classes={
-    DrsStandaloneServer.class,
-    DrsStandaloneSpringConfig.class,
-    StarterKitDrsSpringConfig.class,
+    DrsServer.class,
+    DrsServerSpringConfig.class,
     Objects.class
 })
 @WebAppConfiguration
