+++ src/test/java/org/ga4gh/starterkit/drs/utils/cache/AccessCacheTest.java
@@ -1,7 +1,6 @@
 package org.ga4gh.starterkit.drs.utils.cache;
 
-import org.ga4gh.starterkit.drs.app.DrsStandaloneSpringConfig;
-import org.ga4gh.starterkit.drs.beanconfig.StarterKitDrsSpringConfig;
+import org.ga4gh.starterkit.drs.app.DrsServerSpringConfig;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ContextConfiguration;
@@ -13,8 +12,7 @@ import org.testng.annotations.Test;
 
 @SpringBootTest
 @ContextConfiguration(classes={
-    StarterKitDrsSpringConfig.class,
-    DrsStandaloneSpringConfig.class
+    DrsServerSpringConfig.class
 })
 @WebAppConfiguration
 public class AccessCacheTest extends AbstractTestNGSpringContextTests {
